backend/requirements.txt
# ─── FastAPI & Server ───────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# ─── Data & ML ──────────────────────────────────────
numpy>=1.24.0
joblib>=1.3.0
torch>=2.0.0
torchaudio>=2.0.0
speechbrain>=1.0.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# ─── Database ───────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ─── Utilities ──────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0
